requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pandas
openpyxl
python-dotenv
jira
sqlalchemy
psycopg2-binary
python-multipart
openai>=1.0.0
jinja2>=3.1.2
psycopg2-binary
sqlalchemy